fix and add credit
