+++ NOTICE.txt
@@ -6,3 +6,8 @@ The Apache Software Foundation (http://www.apache.org/).
 
 This product includes software from the twox-hash project (MIT License)
 https://github.com/shepmaster/twox-hash
+
+This product includes software developed at
+Apache Gluten (https://github.com/apache/incubator-gluten/)
+Specifically:
+- Optimizer rule to replace SortMergeJoin with ShuffleHashJoin
